Move `home` tests to a package.
